+++ packages/app-account/src/interbit/control/actions.js
@@ -1,15 +1,6 @@
-const covenantName = 'app-account-control'
+const actionTypes = {}
 
-const actionTypes = {
-  PRIVATE_CHAIN_COVENANT: `${covenantName}/PRIVATE_CHAIN_COVENANT`
-}
-
-const actionCreators = {
-  privateChainCovenant: ({ chainAlias, covenantHash }) => ({
-    type: actionTypes.PRIVATE_CHAIN_COVENANT,
-    payload: { chainAlias, covenantHash }
-  })
-}
+const actionCreators = {}
 
 module.exports = {
   actionTypes,