+++ packages/app-account/src/interbit/my-account/index.js
@@ -61,6 +61,9 @@ const reducer = (state = initialState, action) => {
         joinName
       })
 
+      // HACK: Same permissions as the current action
+      provideAction.publicKey = action.publicKey
+
       console.log('REDISPATCH: ', provideAction)
       nextState = redispatch(nextState, provideAction)
 
@@ -118,6 +121,9 @@ const reducer = (state = initialState, action) => {
         mount: [...PATHS.PRIVATE_PROFILE, tokenName],
         joinName
       })
+
+      // HACK: Same permissions as the current action
+      consumeAction.publicKey = action.publicKey
 
       console.log('REDISPATCH: ', consumeAction)
       nextState = redispatch(nextState, consumeAction)