+++ packages/app-interbit-io/src/constants/index.js
@@ -2,15 +2,23 @@ import ga from './ga'
 import layout from './layout'
 import urls from './urls'
 import paths from './paths'
-import { headerNav, footerNav, footerBottomLinks } from './navigation'
+import pipedrive from './pipedrive'
+import {
+  headerNav,
+  headerRightNav,
+  footerNav,
+  footerBottomLinks
+} from './navigation'
 
 export default {
   ga,
   layout,
   urls,
   paths,
+  pipedrive,
   navigation: {
     headerNav,
+    headerRightNav,
     footerNav,
     footerBottomLinks
   }