+++ packages/app-interbit-io/src/constants/navigation.js
@@ -47,6 +47,10 @@ export const footerNav = [
         to: paths.DEVELOPERS_EXAMPLES
       },
       {
+        text: 'Docs',
+        to: urls.SDK_DOCS
+      },
+      {
         text: 'Support',
         to: paths.DEVELOPERS_SUPPORT
       }
@@ -104,7 +108,28 @@ export const footerBottomLinks = [
 ]
 
 export const headerNav = [
-  { title: 'Platform', eventKey: 'platform' },
-  { title: 'Developers', eventKey: 'developers' },
-  { title: 'Pricing', eventKey: 'pricing' }
+  {
+    text: 'Platform',
+    to: paths.PLATFORM,
+    eventKey: 'platform'
+  },
+  {
+    text: 'Developers',
+    to: paths.DEVELOPERS,
+    eventKey: 'developers'
+  },
+  {
+    text: 'Pricing',
+    to: paths.PRICING,
+    eventKey: 'pricing'
+  }
+]
+
+export const headerRightNav = [
+  {
+    text: 'Go to Accounts',
+    to: urls.APP_ACCOUNT,
+    eventKey: 'accounts',
+    className: 'accounts-link'
+  }
 ]