+++ packages/app-interbit-io/src/containers/Contact.js
@@ -1,32 +1,30 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { Row, Col } from 'react-bootstrap'
-import { Markdown } from 'lib-react-interbit'
+import { Markdown } from 'interbit-ui-components'
 
-import Navigation from '../components/Navigation'
 import layout from '../constants/layout'
+import pipedrive from '../constants/pipedrive'
 
 const mapStateToProps = state => ({
   content: state.content.contact
 })
 
 class Contact extends Component {
-  componentDidMount() {
-    const addScript = document.createElement('script')
-    addScript.setAttribute(
-      'src',
-      'https://webforms.pipedriveassets.com/webforms.min.js'
-    )
-    document.body.appendChild(addScript)
-  }
-
   render() {
     // eslint-disable-next-line react/prop-types
     const { content } = this.props
     const colLayout = layout.colLayout.default
 
-    const contactContent = (
-      <div className="ibweb-page">
+    const uuid = Math.random()
+      .toString(36)
+      .substring(7)
+    const iframeSrc = `https://pipedrivewebforms.com/form/${
+      pipedrive.CONTACT_FORM_ID
+    }?embeded=1&uudid=${uuid}`
+
+    return (
+      <div className="ibweb-page contact">
         <Row>
           <Col {...colLayout}>
             <h1>{content.title}</h1>
@@ -39,19 +37,16 @@ class Contact extends Component {
             <div className="pipedrive-container">
               <div
                 className="pipedriveWebForms"
-                data-pd-webforms="https://pipedrivewebforms.com/form/7148955b25f0e873fca3d9eb37c1118f906729"
-              />
+                data-pd-webforms={`https://pipedrivewebforms.com/form/${
+                  pipedrive.CONTACT_FORM_ID
+                }`}
+                data-uuid={uuid}>
+                <iframe title="ib-pipedrive-form" src={iframeSrc} />
+              </div>
             </div>
           </Col>
         </Row>
       </div>
-    )
-
-    return (
-      <Navigation
-        container={contactContent}
-        className="app-interbit-io contact"
-      />
     )
   }
 }