+++ packages/app-sdk-incrementer/src/index.js
@@ -5,10 +5,10 @@ import { Provider } from 'react-redux'
 import { createStore, applyMiddleware } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import { BrowserRouter } from 'react-router-dom'
-import { middleware as interbitMiddleware } from 'interbit-middleware'
+import { middleware as interbitMiddleware } from 'interbit-ui-tools'
 
-import 'lib-react-interbit/src/css/index.css'
-import 'lib-react-interbit/src/css/interbit.css'
+import 'interbit-ui-components/src/css/index.css'
+import 'interbit-ui-components/src/css/interbit.css'
 
 import App from './App'