+++ packages/interbit-ui-tools/src/index.js
@@ -1,0 +1,20 @@
+// © 2018 BTL GROUP LTD -  This package is licensed under the MIT license https://opensource.org/licenses/MIT
+const { actionTypes, actionCreators, chainDispatch } = require('./actions')
+const createMiddleware = require('./middleware')
+const reducer = require('./reducer')
+const { rootSaga } = require('./sagas')
+const selectors = require('./selectors')
+const { INTERBIT_REDUCER_KEY } = require('./constants')
+const cAuthRequests = require('./cAuth')
+
+module.exports = {
+  actionTypes,
+  actionCreators,
+  chainDispatch,
+  createMiddleware,
+  reducer,
+  rootSaga,
+  selectors,
+  INTERBIT_REDUCER_KEY,
+  cAuthRequests
+}