+++ packages/app-interbit-io/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.17"></a>
+## [0.4.17](https://github.com/interbit/interbit/compare/v0.4.16...v0.4.17) (2018-05-08)
+
+
+
+
+**Note:** Version bump only for package app-interbit-io
+
+<a name="0.4.16"></a>
+## [0.4.16](https://github.com/interbit/interbit/compare/v0.4.15...v0.4.16) (2018-05-07)
+
+
+### Bug Fixes
+
+* Pipedrive form now loads when Contact page accessed not on initial load of app ([8ccf523](https://github.com/interbit/interbit/commit/8ccf523))
+
+
+
+
+<a name="0.4.15"></a>
+## [0.4.15](https://github.com/interbit/interbit/compare/v0.4.14...v0.4.15) (2018-05-02)
+
+
+
+
+**Note:** Version bump only for package app-interbit-io
+
 <a name="0.4.14"></a>
 ## [0.4.14](https://github.com/interbit/interbit/compare/v0.4.13...v0.4.14) (2018-04-27)